+++ history.md
@@ -1,3 +1,8 @@
+# 0.2.0 (2019-11-09)
+* allow to override filename
+* add black and white method lists
+* add customizable encode/decode routines
+
 # 0.1.1 (2019-11-08)
 * add functions support